Validated two blocks model until fracture is fully sliding
The active set algorithm exhibits instabilities when the crack fully
slides. This happen if the bottom line of elements (which are
constrained in the vertical direction) is allowed to change state.
A possible way to stabilize the problem is forcing the bottom line of
elements to stay in stick mode, and in general, forcing any fracture
element that touch the boundary to stay in stick mode.