Updated BC dof removal for Mortar slave dirichlet nodes (#19)
This PR modifies the removal of Dirichlet nodes in Mortar slave side, that is needed for nodal Multipliers.
Now, the BC entities are removed once forever in the constructor of the Mortar object.
A new method removeBCentities() in the Boundaries class has been introduced to remove BC entities (or loaded entities) after the BoundaryConditions have already been defined.